pyproject.toml
[tool.poetry]
name = "fast_api_playground"
version = "0.1.0"
description = "Playground for experimenting with FastAPI"
authors = ["srholand"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
aiohttp = "^3.8.1"
SQLAlchemy = "^1.4.36"
alembic = "^1.7.7"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
